requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
pgvector==0.3.2
redis[hiredis]==5.0.7
google-generativeai==0.7.2
tiktoken==0.7.0
pydantic-settings==2.3.4
python-multipart==0.0.9
openai>=1.60.0
